requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.1.0
pydantic>=2.10.0

# LangGraph + LLM
langgraph>=0.4.0
langchain-openai>=0.3.0
langchain-core>=0.3.0

# Repo cloning
gitpython>=3.1.44

# Report generation
markdown>=3.7
weasyprint>=63.0
pygments>=2.19.0

# Frontend
streamlit>=1.42.0
requests>=2.32.0

# Utilities
tiktoken>=0.9.0
